requirements.txt
mlflow
dvc
dagshub
pandas
scikit-learn
pyyaml
pathlib
numpy
ipykernel
python-dotenv
pymongo
fastapi
uvicorn
pymongo[srv]
requests
dataclasses
python-box